requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.23.0
pydantic>=2.11.7,<3.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
